Move geo functions from index.js to geo.js
